Consumer Service/pyproject.toml
[project]
name = "consumer-service"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "boto3>=1.42.57",
    "pydantic-settings>=2.13.1",
    "python-dotenv>=1.2.1",
]
